[PERF-274] Add NewRelic instrumentation to contentserver.
This includes the course key/path for a given contentserver request,
whether or not the asset is locked, cacheable, has been modified, etc.

This should give us some better insight as to what sort of requests are
coming into the contentserver to figure out whether or use of a CDN in
front of these assets is as efficient as effective as it could be.